backend/requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
pymysql==1.1.0
flask-jwt-extended==4.6.0
werkzeug==3.0.1
python-barcode==0.15.1
Pillow==10.1.0
python-dotenv==1.0.0
python-dateutil==2.8.2
cloudinary==1.36.0
gunicorn==21.2.0